fix: move to vite
