--- .env
@@ -1,1 +1,0 @@
-DISABLE_NEW_JSX_TRANSFORM="true"
