+++ .env.example
@@ -1,2 +1,1 @@
-DISABLE_NEW_JSX_TRANSFORM="true"
-REACT_APP_MIXPANEL_TOKEN=""
+VITE_MIXPANEL_TOKEN=""
