+++ .eslintignore
@@ -1,1 +1,3 @@
+*.svg
+*.png
 *.html
