+++ .eslintrc.json
@@ -1,34 +1,17 @@
 {
-  "parser": "@typescript-eslint/parser",
-  "parserOptions": {
-    "ecmaVersion": 2020,
-    "sourceType": "module",
-    "ecmaFeatures": {
-      // Allows for the parsing of JSX
-      "jsx": true
-    }
-  },
-  "ignorePatterns": ["node_modules/**/*"],
-  "settings": {
-    "react": {
-      "version": "detect"
-    }
-  },
+  "env": { "browser": true, "es2020": true },
   "extends": [
-    "plugin:react/recommended",
+    "eslint:recommended",
     "plugin:@typescript-eslint/recommended",
     "plugin:react-hooks/recommended",
     "prettier/@typescript-eslint",
     "plugin:prettier/recommended"
   ],
-  "plugins": ["react", "react-hooks", "prettier"],
+  "parser": "@typescript-eslint/parser",
+  "parserOptions": { "ecmaVersion": "latest", "sourceType": "module" },
+  "plugins": ["react-refresh", "prettier"],
   "rules": {
-    "react/jsx-filename-extension": [
-      1,
-      {
-        "extensions": [".js", ".jsx", ".ts", ".tsx"]
-      }
-    ],
+    "react-refresh/only-export-components": "off",
     "@typescript-eslint/explicit-function-return-type": "off",
     "prettier/prettier": "error",
     "react/jsx-uses-react": "off",
