+++ index.html
@@ -1,0 +1,58 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <link rel="shortcut icon" type="image/png" href="/favicon.png" id="favicon" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <meta name="theme-color" content="#000000" />
+    <meta name="description" content="Pangolin - Compound Your Community" />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Poppins:wght@100;200;300;400;500;600;700;800;900&display=swap"
+      rel="stylesheet"
+    />
+    <title>Pangolin - Compound Your Community</title>
+  </head>
+  <body>
+    <div id="root"></div>
+    <script>window.global = window;</script>
+    <script type="module" src="/src/index.tsx"></script>
+    <script>
+      window.intercomSettings = {
+        app_id: 'uj2kwt4z'
+      }
+    </script>
+    <script>
+      ;(function() {
+        var w = window
+        var ic = w.Intercom
+        if (typeof ic === 'function') {
+          ic('reattach_activator')
+          ic('update', w.intercomSettings)
+        } else {
+          var d = document
+          var i = function() {
+            i.c(arguments)
+          }
+          i.q = []
+          i.c = function(args) {
+            i.q.push(args)
+          }
+          w.Intercom = i
+          var l = function() {
+            var s = d.createElement('script')
+            s.type = 'text/javascript'
+            s.async = true
+            s.src = 'https://widget.intercom.io/widget/uj2kwt4z'
+            var x = d.getElementsByTagName('script')[0]
+            x.parentNode.insertBefore(s, x)
+          }
+          if (w.attachEvent) {
+            w.attachEvent('onload', l)
+          } else {
+            w.addEventListener('load', l, false)
+          }
+        }
+      })()
+    </script>
+  </body>
+</html>
