+++ package.json
@@ -1,41 +1,28 @@
 {
   "name": "pangolin.exchange",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
-  "devDependencies": {
-    "@types/jest": "27.5.2",
-    "@types/mixpanel-browser": "2.38.0",
-    "@types/node": "16.11.58",
-    "@types/qs": "6.9.7",
-    "@types/react": "16.14.24",
-    "@types/react-dom": "16.9.14",
-    "@types/react-redux": "7.1.23",
-    "@types/styled-components": "5.1.24",
-    "@typescript-eslint/eslint-plugin": "4.6.0",
-    "@typescript-eslint/parser": "4.6.0",
-    "babel-jest": "26.6.3",
-    "babel-plugin-styled-components": "2.0.6",
-    "@babel/plugin-proposal-optional-chaining": "7.18.9",
-    "eslint": "7.11.0",
-    "eslint-config-prettier": "6.15.0",
-    "eslint-plugin-prettier": "3.4.1",
-    "eslint-plugin-react": "7.29.4",
-    "eslint-plugin-react-hooks": "4.3.0",
-    "husky": "7.0.4",
-    "lint-staged": "12.3.7",
-    "npm-force-resolutions": "0.0.10",
-    "patch-package": "6.4.7",
-    "prettier": "1.19.1"
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "lint": "eslint src --ext ts,tsx --report-unused-disable-directives",
+    "lint:fix": "eslint src --ext ts,tsx --report-unused-disable-directives --fix",
+    "preview": "vite preview",
+    "prepare": "husky install",
+    "type-check": "tsc --noEmit",
+    "pre-commit-hook": "lint-staged"
   },
   "dependencies": {
-    "@babel/plugin-proposal-nullish-coalescing-operator": "7.18.6",
-    "@babel/core": "7.21.8",
     "@pangolindex/components": "7.1.3",
     "@pangolindex/sdk": "5.1.0",
     "@reduxjs/toolkit": "1.8.0",
     "@testing-library/jest-dom": "5.16.5",
     "@testing-library/react": "12.1.4",
     "@testing-library/user-event": "13.5.0",
+    "@web3-react/core": "6.0.9",
+    "@web3-react/walletconnect-connector": "6.2.13",
+    "@web3-react/walletlink-connector": "6.2.11",
     "axios": "1.1.3",
     "jest": "26.6.0",
     "mixpanel-browser": "2.45.0",
@@ -45,33 +32,44 @@
     "react-query": "3.39.2",
     "react-redux": "7.2.6",
     "react-router-dom": "6.8.1",
-    "@web3-react/core": "6.0.9",
-    "@web3-react/walletconnect-connector": "6.2.13",
-    "@web3-react/walletlink-connector": "6.2.11",
     "react-scripts": "4.0.3",
     "react-spring": "8.0.27",
     "react-use": "15.3.8",
     "redux-localstorage-simple": "2.4.1",
     "styled-components": "5.3.1",
-    "typescript": "4.9.5",
     "web-vitals": "2.1.4"
   },
-  "resolutions": {
+  "devDependencies": {
+    "@babel/plugin-proposal-optional-chaining": "7.18.9",
+    "@esbuild-plugins/node-globals-polyfill": "0.2.3",
+    "@types/jest": "27.5.2",
+    "@types/mixpanel-browser": "2.38.0",
+    "@types/node": "16.11.58",
+    "@types/qs": "6.9.7",
     "@types/react": "16.14.24",
-    "styled-components": "5.3.1"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject",
-    "version": "npm run build",
-    "lint": "eslint ./src --ext .js,.jsx,.ts,.tsx --fix",
-    "get:version": "node -pe \"require('./package.json').version\"",
-    "prepare": "husky install",
-    "type-check": "tsc --noEmit",
-    "pre-commit-hook": "lint-staged",
-    "postinstall": "patch-package"
+    "@types/react-dom": "16.9.14",
+    "@types/react-redux": "7.1.23",
+    "@types/styled-components": "5.1.24",
+    "@typescript-eslint/eslint-plugin": "^5.59.0",
+    "@typescript-eslint/parser": "^5.59.0",
+    "@vitejs/plugin-react": "^4.0.0",
+    "babel-jest": "26.6.3",
+    "babel-plugin-styled-components": "2.0.6",
+    "eslint": "^8.38.0",
+    "eslint-config-prettier": "6.15.0",
+    "eslint-plugin-prettier": "3.4.1",
+    "eslint-plugin-react": "7.29.4",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.3.4",
+    "husky": "7.0.4",
+    "lint-staged": "12.3.7",
+    "npm-force-resolutions": "0.0.10",
+    "patch-package": "6.4.7",
+    "prettier": "1.19.1",
+    "typescript": "^5.0.2",
+    "vite": "^4.3.9",
+    "vite-plugin-svgr": "3.2.0",
+    "vite-tsconfig-paths": "4.2.0"
   },
   "lint-staged": {
     "*/!(node_modules|dist)|**/**/*.{jsx,js,ts,tsx}": [
@@ -80,16 +78,8 @@
       "git add ."
     ]
   },
-  "eslintConfig": {
-    "extends": "react-app",
-    "ignorePatterns": [
-      "node_modules"
-    ]
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not op_mini all"
-  ],
-  "license": "MIT"
+  "resolutions": {
+    "@types/react": "16.14.24",
+    "styled-components": "5.3.1"
+  }
 }
