--- public/index.html
@@ -1,81 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <link rel="shortcut icon" type="image/png" href="%PUBLIC_URL%/favicon.png" id="favicon" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#000000" />
-    <meta name="description" content="Pangolin - Compound Your Community" />
-    <!--
-      manifest.json provides metadata used when your web app is installed on a
-      user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
-    -->
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Poppins:wght@100;200;300;400;500;600;700;800;900&display=swap"
-      rel="stylesheet"
-    />
-    <!--
-      Notice the use of %PUBLIC_URL% in the tags above.
-      It will be replaced with the URL of the `public` folder during the build.
-      Only files inside the `public` folder can be referenced from the HTML.
-
-      Unlike "/favicon.ico" or "favicon.ico", "%PUBLIC_URL%/favicon.ico" will
-      work correctly both with client-side routing and a non-root public URL.
-      Learn how to configure a non-root public URL by running `npm run build`.
-    -->
-    <title>Pangolin - Compound Your Community</title>
-  </head>
-  <body>
-    <script>
-      window.intercomSettings = {
-        app_id: 'uj2kwt4z'
-      }
-    </script>
-    <script>
-      ;(function() {
-        var w = window
-        var ic = w.Intercom
-        if (typeof ic === 'function') {
-          ic('reattach_activator')
-          ic('update', w.intercomSettings)
-        } else {
-          var d = document
-          var i = function() {
-            i.c(arguments)
-          }
-          i.q = []
-          i.c = function(args) {
-            i.q.push(args)
-          }
-          w.Intercom = i
-          var l = function() {
-            var s = d.createElement('script')
-            s.type = 'text/javascript'
-            s.async = true
-            s.src = 'https://widget.intercom.io/widget/uj2kwt4z'
-            var x = d.getElementsByTagName('script')[0]
-            x.parentNode.insertBefore(s, x)
-          }
-          if (w.attachEvent) {
-            w.attachEvent('onload', l)
-          } else {
-            w.addEventListener('load', l, false)
-          }
-        }
-      })()
-    </script>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
-    <!--
-      This HTML file is a template.
-      If you open it directly in the browser, you will see an empty page.
-
-      You can add webfonts, meta tags, or analytics to this file.
-      The build step will place the bundled scripts into the <body> tag.
-
-      To begin the development, run `npm start` or `yarn start`.
-      To create a production bundle, use `npm run build` or `yarn build`.
-    -->
-  </body>
-</html>
